Refactor dependencies for `#![no_std]` compatibility
Three of the crates depended upon ― `byteorder`, `num_enum`, and `strum` respectively ― need to have `default-features = false` applied in order for this crate to actually be `#![no_std]` compatible. These dependency crates try to pull in `std` ― breaking any usage of this crate in `#![no_std]` environments by so doing ― if this is not specified.